# api/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "FieldViolation": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Message": {
        "description": "Application type Message.",
        "type": "object"
      },
      "MessageList": {
        "description": "Application type MessageList.",
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "code": {
            "type": "string"
          },
          "currentRevision": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "fieldViolations": {
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            },
            "type": "array"
          },
          "instance": {
            "type": "string"
          },
          "remediation": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "status": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "type": "object"
      },
...
      "WaitIn": {
        "description": "Application type WaitIn.",
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "token",
        "description": "Lab static bearer. Auth is stubbed open in API-001; SEC-001 adds bearer + basic.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST adapter for the shared LabMail capability registry. Generated from internal/capabilities. Do not edit by hand.",
    "title": "LabMail Management API",
    "version": "v1",
    "x-generated-by": "internal/control/rest.RenderOpenAPI; DO NOT EDIT."
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/messages": {
      "delete": {
        "description": "Delete every message. Does not bump epoch.",
        "operationId": "delete_v1_messages",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Clear inbox"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "application/problem+json domain error"
          }
        },
        "summary": "Clear inbox",
        "tags": [
          "messages"
        ],
        "x-capability-id": "messages.clear",
        "x-idempotent": true,
        "x-mutating": true,
        "x-required-scopes": [
          "mail.write"
        ]
      },
      "get": {
        "description": "Cursor-paginated inbox list. HMAC cursors bind storeGeneration.",
        "operationId": "get_v1_messages",
        "parameters": [
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "subject",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "subjectContains",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "unread",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "after",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "before",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageList"
                }
              }
            },
            "description": "List messages"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "application/problem+json domain error"
          }
        },
        "summary": "List messages",
        "tags": [
          "messages"
        ],
        "x-capability-id": "messages.list",
        "x-idempotent": true,
        "x-mutating": false,
        "x-required-scopes": [
          "mail.read"
        ]
      }
    },
    "/v1/messages/{id}": {
      "delete": {
        "description": "Delete one message by id.",
        "operationId": "delete_v1_messages_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Delete message"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "application/problem+json domain error"
          }
        },
        "summary": "Delete message",
        "tags": [
          "messages"
        ],
        "x-capability-id": "messages.delete",
        "x-idempotent": true,
        "x-mutating": true,
        "x-required-scopes": [
          "mail.write"
        ]
      },
      "get": {
        "description": "Full message. markRead defaults to false on the native path.",
        "operationId": "get_v1_messages_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Native default is false. Compat GET /email/:id marks read.",
            "in": "query",
            "name": "markRead",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Get message"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "application/problem+json domain error"
          }
        },
        "summary": "Get message",
        "tags": [
          "messages"
        ],
        "x-capability-id": "messages.get",
        "x-idempotent": true,
        "x-mutating": false,
        "x-required-scopes": [
          "mail.read"
        ]
      }
    },
...
    "/v1/messages:wait": {
      "post": {
        "description": "Block until a matching message arrives or the timeout fires.",
        "operationId": "post_v1_messages_wait",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WaitIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Wait for message"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "application/problem+json domain error"
          }
        },
        "summary": "Wait for message",
        "tags": [
          "messages"
        ],
        "x-capability-id": "messages.wait",
        "x-idempotent": true,
        "x-mutating": false,
        "x-required-scopes": [
          "mail.read"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Management listener (default address :1080)",
      "url": "/"
    }
  ],
...
}